+++ guillotina_audit/__init__.py
@@ -14,7 +14,8 @@ app_settings = {
         }
     },
     "commands": {
-        "audit-update-mappings": "guillotina_audit.commands.mappings.UpdateMappingsCommand"
+        "audit-update-mappings": "guillotina_audit.commands.mappings.UpdateMappingsCommand",
+        "audit-update-settings": "guillotina_audit.commands.settings.UpdateSettingsCommand",
     },
 }